requirements.txt
asgiref==3.2.3
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.8
cx-Oracle==7.3.0
defusedxml==0.6.0
Django==2.1.1
django-allauth==0.41.0
django-cors-headers==3.2.1
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
djangorestframework==3.11.0
et-xmlfile==1.0.1
gunicorn==20.0.4
http-ece==1.1.0
idna==2.8
itypes==1.1.0
jdcal==1.4.1
Jinja2==2.10.3
MarkupSafe==1.1.1
mysqlclient==1.3.12
oauthlib==3.1.0
openapi-codec==1.3.2
openpyxl==3.0.3
Pillow==7.0.0
psycopg2==2.8.4
py-vapid==1.7.0
pycparser==2.19
pymssql==2.1.4
python3-openid==3.1.0
pytz==2019.3
pywebpush==1.10.1
requests==2.22.0
requests-oauthlib==1.3.0
selenium==3.141.0
simplejson==3.17.0
six==1.14.0
sqlparse==0.3.0
uritemplate==3.0.1
urllib3==1.25.7
